requirements.txt
torch
torchvision
transformers
ultralytics
pillow
langchain
langchain-core
langgraph
langchain-groq
python-dotenv
gradio
opencv-python
numpy

